+++ app/src/main/kotlin/space/zuoqiu/siyuanhelper/data/Subject.kt
@@ -25,6 +25,7 @@ data class Subject(
     val id: Long, // 主题 ID
     var title: String, // 主题标题
     var description: String, // 主题描述
+    var isDefault: Boolean, // 是否设置为默认主题
     var updated: ZonedDateTime = ZonedDateTime.now(), // 更新时间
 ) {
 }