+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - 实现基本 UI 框架 | Implement basic UI framework
   - REF: https://github.com/android/compose-samples/tree/main/Reply
 - 实现主题编辑功能 | Implement subject editing function
+- 支持主题钉住与备忘录星标功能 | Support pin subject and star memo function